+++ src/Stempler/src/Directive/DirectiveGroup.php
@@ -1,12 +1,5 @@
 <?php
 
-/**
- * Spiral Framework.
- *
- * @license   MIT
- * @author    Anton Titov (Wolfy-J)
- */
-
 declare(strict_types=1);
 
 namespace Spiral\Stempler\Directive;
@@ -15,12 +8,12 @@ use Spiral\Stempler\Node\Dynamic\Directive;
 
 final class DirectiveGroup implements DirectiveRendererInterface
 {
-    /** @var DirectiveRendererInterface[] */
-    private $directives = [];
-
-    public function __construct(array $directives = [])
-    {
-        $this->directives = $directives;
+    /**
+     * @param \Spiral\Stempler\Directive\DirectiveRendererInterface[] $directives
+     */
+    public function __construct(
+        private array $directives = []
+    ) {
     }
 
     /**