+++ src/Stempler/src/Parser/Context.php
@@ -1,12 +1,5 @@
 <?php
 
-/**
- * Spiral Framework.
- *
- * @license   MIT
- * @author    Anton Titov (Wolfy-J)
- */
-
 declare(strict_types=1);
 
 namespace Spiral\Stempler\Parser;
@@ -18,23 +11,13 @@ use Spiral\Stempler\Lexer\Token;
  */
 final class Context
 {
-    public $parent;
-    /** @var Token */
-    private $token;
+    public ?Context $parent = null;
+    private array $values = [];
 
-    /** @var string|null */
-    private $path;
-
-    /** @var array */
-    private $values = [];
-
-    /**
-     * @param string|null $path
-     */
-    public function __construct(Token $token, string $path = null)
-    {
-        $this->token = $token;
-        $this->path = $path;
+    public function __construct(
+        private Token $token,
+        private ?string $path = null
+    ) {
     }
 
     public function getToken(): Token
@@ -47,21 +30,15 @@ final class Context
         return $this->path;
     }
 
-    /**
-     * @param mixed  $value
-     */
-    public function withValue(string $name, $value): self
+    public function withValue(string $name, mixed $value): self
     {
         $ctx = clone $this;
         $ctx->values[$name] = $value;
+
         return $ctx;
     }
 
-    /**
-     * @param null   $default
-     * @return mixed|null
-     */
-    public function getValue(string $name, $default = null)
+    public function getValue(string $name, mixed $default = null): mixed
     {
         return $this->values[$name] ?? $default;
     }