+++ src/Stempler/src/VisitorInterface.php
@@ -1,17 +1,8 @@
 <?php
 
-/**
- * Spiral Framework.
- *
- * @license   MIT
- * @author    Anton Titov (Wolfy-J)
- */
-
 declare(strict_types=1);
 
 namespace Spiral\Stempler;
-
-use Spiral\Stempler\Node\NodeInterface;
 
 interface VisitorInterface
 {
@@ -50,17 +41,7 @@ interface VisitorInterface
      */
     public const DONT_TRAVERSE_CURRENT_AND_CHILDREN = 4;
 
-    /**
-     * @param NodeInterface|mixed $node
-     * @param VisitorContext      $ctx
-     * @return mixed
-     */
-    public function enterNode($node, VisitorContext $ctx);
+    public function enterNode(mixed $node, VisitorContext $ctx): mixed;
 
-    /**
-     * @param NodeInterface|mixed $node
-     * @param VisitorContext      $ctx
-     * @return mixed
-     */
-    public function leaveNode($node, VisitorContext $ctx);
+    public function leaveNode(mixed $node, VisitorContext $ctx): mixed;
 }